split export code
