+++ packages/excalidraw/scene/export.ts
@@ -11,7 +11,7 @@ import {
   getCommonBounds,
   getElementAbsoluteCoords,
 } from "../element/bounds";
-import { renderSceneToSvg } from "../renderer/renderScene";
+import { renderSceneToSvg } from "../renderer/export";
 import { arrayToMap, distance, getFontString, toBrandedType } from "../utils";
 import { AppState, BinaryFiles } from "../types";
 import {
